Feature: Warning for high or low max allocated RAM
Added a method which checks the ram allocation and then gives a warning if it is offset by 1100 MB from the recommended max ram allocation. The offset for 1100 is set because AA-like categories use about 1000 MB more than other categories. MultiMC's Ram Allocation amounts may pass the 1000 MB barrier, to avoid it we add 100 more to the 1000 MB. There's also a system property to disable it.